Split env var substitution from other substitutions
Use that from the file provider
